The Iris dataset
The following table depicts an excerpt of the Iris dataset, which is a classic example in the field of machine learning. The Iris dataset contains the measurements of 150 Iris flowers
from three different species—Setosa, Versicolor, and Virginica. Here, each flower
sample represents one row in our dataset, and the flower measurements in centimeters are stored as columns, which we also call the features of the dataset: